Use 32-bit ghost IDs; bump protocol
Switch player/ghost identifiers from 8-bit to 32-bit across client, server and protocol; bump Protocol.Version to 7 and add Protocol.GhostIdLen=4. Update wire framing, payload lengths, parsing and builders accordingly (many GameBridge, ClientSession, InteractionClient, VoiceChat, TcpBroadcastService, VersionIpcServer, launcher models, etc.). Add AsyncLogger (non-blocking console output + summaries) and wire it into GameBridge. Minor validator/message polish in AnnounceValidator. These changes enable larger player counts and avoid id-space limits.

+++ KCDMP_launcher/Models/AppModels.cs
@@ -40,7 +40,7 @@ namespace KCDMP_launcher.Models
 
     public class PeerVersionData
     {
-        public byte GhostId { get; set; }
+        public uint GhostId { get; set; }
         public string ReleaseVersion { get; set; } = "";
     }
